move tables with build constraints
